+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- fix toolbar and toolbar button components to avoid creating them twice ([#1](https://github.com/seaofvoices/react-roblox-studio-plugin/pull/1))
+
 ## 0.1.0
 
 - Initial version